refactor: 렌주룰과 오목 판정 로직 이동
