+++ src/main/kotlin/omok/domain/stone/BlackStones.kt
@@ -1,14 +1,12 @@
 package omok.domain.stone
 
-import rule.BlackRenjuRule
-import rule.OmokRule
-import rule.type.Violation
-import rule.wrapper.point.Point
+import omok.domain.Point
+import omok.domain.rule.RenjuRule
 
 class BlackStones(
     points: Set<Point> = emptySet(),
 ) : Stones(points) {
-    override val rule: OmokRule = BlackRenjuRule()
+    private val renjuRule = RenjuRule(points)
 
     override fun plus(point: Point): BlackStones = BlackStones(points + point)
 
@@ -16,8 +14,6 @@ class BlackStones(
         other: Stones,
         point: Point,
     ): Boolean {
-        val violateType =
-            rule.checkAnyFoulCondition(this.points.toList(), other.points.toList(), point)
-        return violateType.state
+        return renjuRule.isFoul(other, point)
     }
 }
