+++ src/main/kotlin/omok/domain/stone/Stones.kt
@@ -1,15 +1,14 @@
 package omok.domain.stone
 
-import rule.OmokRule
-import rule.wrapper.point.Point
+import omok.domain.Point
+import omok.domain.rule.OmokRule
 
 abstract class Stones(
     points: Set<Point> = emptySet(),
 ) {
+    private val omokRule = OmokRule(points)
     private val _points = points.toMutableSet()
     val points = _points.toSet()
-
-    abstract val rule: OmokRule
 
     abstract operator fun plus(point: Point): Stones
 
@@ -20,41 +19,5 @@ abstract class Stones(
     fun isOmok(
         lastPoint: Point,
         size: Int,
-    ): Boolean =
-        listOf(HORIZONTAL, VERTICAL, DIAGONAL_UP, DIAGONAL_DOWN).any {
-            isSerialOmok(lastPoint, it, size)
-        }
-
-    private fun isSerialOmok(
-        lastPoint: Point,
-        directions: List<Pair<Int, Int>>,
-        size: Int,
-    ): Boolean = directions.sumOf { countConnected(lastPoint, it, size) } >= OMOK_STONE_COUNT - 1
-
-    private fun countConnected(
-        point: Point,
-        direction: Pair<Int, Int>,
-        size: Int,
-    ): Int {
-        val boardRange = 0..<size
-        var count = 0
-
-        val (dx, dy) = direction
-        var (x, y) = point.row + dx to point.col + dy
-        while (x in boardRange && y in boardRange && Point(x, y) in points) {
-            count++
-            x += dx
-            y += dy
-        }
-        return count
-    }
-
-    companion object {
-        private const val OMOK_STONE_COUNT = 5
-
-        private val HORIZONTAL = listOf(Pair(-1, 0), Pair(1, 0))
-        private val VERTICAL = listOf(Pair(0, 1), Pair(0, -1))
-        private val DIAGONAL_UP = listOf(Pair(-1, -1), Pair(1, 1))
-        private val DIAGONAL_DOWN = listOf(Pair(-1, 1), Pair(1, -1))
-    }
+    ): Boolean = omokRule.isOmok(lastPoint, size)
 }
