+++ src/main/kotlin/omok/domain/stone/WhiteStones.kt
@@ -1,13 +1,9 @@
 package omok.domain.stone
 
-import rule.OmokRule
-import rule.WhiteRenjuRule
-import rule.wrapper.point.Point
+import omok.domain.Point
 
 class WhiteStones(
     points: Set<Point> = emptySet(),
 ) : Stones(points) {
-    override val rule: OmokRule = WhiteRenjuRule()
-
     override fun plus(point: Point): WhiteStones = WhiteStones(points + point)
 }
